+++ src/app/projectmanagement/overview.component.ts
@@ -23,6 +23,8 @@ import {NgForm} from '@angular/forms';
 import {Flavor} from '../virtualmachines/virtualmachinemodels/flavor';
 import {FlavorType} from '../virtualmachines/virtualmachinemodels/flavorType';
 import {FlavorService} from '../api-connector/flavor.service';
+import {IResponseTemplate} from '../api-connector/response-template';
+import {CreditsService} from '../api-connector/credits.service';
 
 /**
  * Projectoverview component.
@@ -31,7 +33,7 @@ import {FlavorService} from '../api-connector/flavor.service';
              selector: 'app-project-overview',
              templateUrl: 'overview.component.html',
              providers: [FlavorService, ApplicationStatusService, ApplicationsService,
-               FacilityService, VoService, UserService, GroupService, ApiSettings]
+               FacilityService, VoService, UserService, GroupService, ApiSettings, VoService, CreditsService]
            })
 export class OverviewComponent extends ApplicationBaseClassComponent implements OnInit {
 
@@ -52,6 +54,7 @@ export class OverviewComponent extends ApplicationBaseClassComponent implements 
   application_id: string;
   project: Project;
   application_details_visible: boolean = false;
+  credits: number = 0;
 
   /**
    * id of the extension status.
@@ -93,7 +96,8 @@ export class OverviewComponent extends ApplicationBaseClassComponent implements 
               userservice: UserService,
               private activatedRoute: ActivatedRoute,
               private fullLayout: FullLayoutComponent,
-              private router: Router) {
+              private router: Router,
+              private voservice: VoService, private creditsService: CreditsService) {
     super(userservice, applicationstatusservice, applicationsservice, facilityService);
   }
 
@@ -185,7 +189,19 @@ export class OverviewComponent extends ApplicationBaseClassComponent implements 
     values['project_application_id'] = this.project_application.Id;
     values['total_cores_new'] = this.totalNumberOfCores;
     values['total_ram_new'] = this.totalRAM;
+    values['project_application_renewal_credits'] = this.credits;
+
     this.requestExtension(values);
+
+  }
+
+  /**
+   * Sends a request to the BE to get the initital credits for a new application.
+   */
+  calculateInitialCredits(form: NgForm): void {
+
+    /*todo calculate */
+    this.credits = 0;
 
   }
 
@@ -275,6 +291,7 @@ export class OverviewComponent extends ApplicationBaseClassComponent implements 
       this.getUserinfo();
       this.getListOfFlavors();
       this.getListOfTypes();
+      this.checkVOstatus();
 
     });
 
@@ -532,6 +549,38 @@ export class OverviewComponent extends ApplicationBaseClassComponent implements 
 
   }
 
+  /**
+   * Uses the data from the application form to fill the confirmation-modal with information.
+   * @param form the application form with corresponding data
+   */
+  filterEnteredData(form: NgForm): void {
+    this.generateConstants();
+    this.valuesToConfirm = [];
+    for (const key in form.controls) {
+      if (form.controls[key].value) {
+        if (key === 'project_application_name') {
+          this.projectName = form.controls[key].value;
+          if (this.projectName.length > 50) {
+            this.projectName = `${this.projectName.substring(0, 50)}...`;
+          }
+        }
+        if (key in this.constantStrings) {
+          if (form.controls[key].disabled) {
+            continue;
+          }
+
+          this.valuesToConfirm.push(this.matchString(key.toString(), form.controls[key].value.toString()));
+
+        }
+      }
+
+    }
+    if (!this.project_application_report_allowed) {
+      this.valuesToConfirm.push('Dissemination allowed: No');
+    }
+
+  }
+
   copyToClipboard(text: string): void {
     document.addEventListener('copy', (clipEvent: ClipboardEvent) => {
       clipEvent.clipboardData.setData('text/plain', (text));
@@ -748,5 +797,15 @@ export class OverviewComponent extends ApplicationBaseClassComponent implements 
     } else if (nr_vm === 0) {
       this.min_vm = false;
     }
+  }
+
+  /**
+   * Check vm status.
+   * @param {UserService} userservice
+   */
+  checkVOstatus(): void {
+    this.voservice.isVo().subscribe((result: IResponseTemplate) => {
+      this.is_vo_admin = <boolean><Boolean>result.value;
+    })
   }
 }